Save changed before adding User.class

+++ src/main/java/net/zcoo/seal/repository/UserRepository.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Repository;
 import net.zcoo.seal.model.DAOUser;
 
 @Repository
-public interface UserRepository extends CrudRepository<DAOUser, Integer> {
-    DAOUser findByUsername(String username);
+public interface UserRepository extends CrudRepository<DAOUser, Long> {
+    DAOUser findByUsername(String username); // >>repository get username>>add method to repository>>define repository
+                                             // method
 }
